requirements.txt
numpy>=1.26
scipy>=1.11
pynrrd>=1.0
pytest>=8.0
